MANUAL: migrates GameSetup interface to game-setup-util lib
