+++ libs/spirit-islander/game-setup/util/src/index.ts
@@ -1,2 +1,3 @@
 export * from './lib/create-game-setup';
+export * from './lib/game-setup.interface';
 export * from './lib/spirit-islander-game-setup-util';
