+++ libs/spirit-islander/shared/data-access/src/lib/app-state.service.ts
@@ -4,11 +4,13 @@ import { first, tap } from 'rxjs';
 import { LocalStorageService, State } from '@atocha/core/data-access';
 import {
   Config,
-  GameSetup,
   migrateConfig,
   Options,
 } from '@atocha/spirit-islander/shared/util';
-import { createGameSetup } from '@atocha/spirit-islander/game-setup/util';
+import {
+  GameSetup,
+  createGameSetup,
+} from '@atocha/spirit-islander/game-setup/util';
 
 export interface AppState {
   config: Config;
