+++ libs/spirit-islander/shared/util/src/index.ts
@@ -13,6 +13,5 @@ export * from './lib/game/spirits';
 export * from './lib/app/app-migration';
 export * from './lib/app/combo.interface';
 export * from './lib/app/config.interface';
-export * from './lib/app/game-setup.interface';
 export * from './lib/app/get-valid-combos';
 export * from './lib/app/route.enum';
